filesystem: rename bus1_fs_peer_find_*() to bus1_fs_peer_acquire_*()
Those functions actually acquire a peer, but rather than passing in the
peer handle, you pass in the ID or name. Hence, make sure to keep the
*_acquire_* token in the function name, to highlight the purpose. The
`by_id' and `by_name' suffixes should be sufficient to notice that it is
a lookup function.

While at it, document both functions properly.

Signed-off-by: David Herrmann <dh.herrmann@gmail.com>