filesystem: fix deadlock on domain teardown
We must not hold a write-lock on the domain when draining its peers.
Otherwise, we race peer-lookup during normal operation and might deadlock
on the domain. Rework the active-object to split draining and cleanup and
use it from the filesystem code.

This reworks the whole domain teardown and adds elaborate locking
comments to explain the inner workings of filesystem.c.

Signed-off-by: David Herrmann <dh.herrmann@gmail.com>