filesystem: support unlink() on debug builds
If a debug build is used, we now support unlink(2) on the `bus' file.
This allows extensive tests for full domain teardown on live domains.
Note that this feature is *fully* disabled on normal production builds.
You need to pass BUS1_EXT set to something else than '1' to enable it.

Signed-off-by: David Herrmann <dh.herrmann@gmail.com>