util: add bus1_import_vecs()
New helper function that is a variation of rw_copy_check_uvector() and
its compat-equivalent.

Sadly, the kernel does not export rw_copy_check_uvector, so we need this
helper. Only the much more restricted import_iovec() function is
exported. However, it does not suit are needs, as we need to extend the
iovec array with out own vectors.

Signed-off-by: David Herrmann <dh.herrmann@gmail.com>