queue: add message queues
This adds the implementation of bus1_queue message queues. Each peers has
one of those queues embedded, which then manages all the queued messages
that have not been dequeued by the peer.

The implementation is a straightforward linked-list, with an additional
set of carried file-descriptors that can be installed into the process of
the receiver.

Signed-off-by: David Herrmann <dh.herrmann@gmail.com>