filesystem: implement mmap
Implement mmap() on peer handles. Move most of the code out of the pool.c
file directly into filesystem.c. We don't provide the same locking
guarantees as the other ioctls, so we shouldn't be calling into any of
the basic peer operations. Instead, hard-code everything in filesystem.c
and explain why it is safe.

Signed-off-by: David Herrmann <dh.herrmann@gmail.com>