queue: make a stronger ordering guarantee
The message order should respect causality: if the completion of a queue
or dequeue operation caused the start of another queue operation, the
messages should be ordered accordingly. Even if the operations are on
different peers and the causality is through some side-channel.

Make this clear in the comment defining the global order.

Signed-off-by: Tom Gundersen <teg@jklm.no>