filesystem: make fs_peer->peer rcu-accessible
We want the peer object to be rcu-accessible from within poll()
callbacks, so we have to use rcu-accessors for fs_peer->peer everywhere.
We already guarantee that bus1_peer is freed rcu-delayed, so the only
critical part is to make sure we reset fs_peer->peer to NULL before
calling into bus1_peer_free().

Signed-off-by: David Herrmann <dh.herrmann@gmail.com>