transaction: implement message transactions
This is an initial, untested implementation of message transactions. It
adds a new object, called bus1_transaction, which can be used to transmit
multi-cast message that adhere to our global-order, but at the same time
do *NOT* require a global bus-lock.

This is not hooked up, yet. It is just committed to get an early view at
how things will look like.

Signed-off-by: David Herrmann <dh.herrmann@gmail.com>