filesystem: explain that EFAULT is special
Usually, if we fail for any reason, we always revert anything we did so
far and bail out. However, EFAULT for *OUTPUT PARAMETERS* is special. We
really cannot treat this as fatal error. Instead, the caller screwed up
and we really couldn't care less.

Add a short comment which explicitly states that EFAULT for QUERY does
not cause a revert of the operation.

Signed-off-by: David Herrmann <dh.herrmann@gmail.com>