filesystem: reduce lock contention on domain
The domain lock is the most contended lock on the whole bus. It is used
for all connect/disconnect operations, and *all* peer lookups. Therefore,
so far we used it solely for link/unlink operations and lookups.

However, there was one exception: For simplicity, we called
bus1_peer_free() while holding it as well. Avoid this, as there is really
no necessity for this. Return the stale peer via a context to the caller
stack, and release it after dropping the locks.

Signed-off-by: David Herrmann <dh.herrmann@gmail.com>