transaction: allocate the final sequence number under the peer lock
As we skip staging the message on the first peer, we risk a race where
message A is assigned its final sequence number before message B, but B
gets queued and dequeued on the 'first' peer before A is queued.

To avoid doing the atomic operation under the lock, we would need to
revert to staging all messages, as we originally did.

Signed-off-by: Tom Gundersen <teg@jklm.no>