transaction: do not assign sequence numbers for unicast messages
Currently, any message is globaly ordered against any other message.
It is indeed necessary (unless we fundamentally change the logic) to
globaly order any message against any multicast message, or we risk
not respecting causality. Consider the peer A sending the (multicast)
message X to peers B and C, and when C receives X it triggers (via
a sidechannel) the peer D to send the message Y to B. We now need the
unicast message Y to be ordered after the multicast message X, even
though from the point of view of the bus they are independent.

However, we could relax the ordering guarantee when it comes to
unicast messages. Consider peer A delivering the messages X and Y
to the peers B and C respectively, and consdir them to be equally
ordered against all multicast messages. Whether or not we
internally consider X to be ordered before Y or Y before X will
not make a difference to how they are delivered (given A and B being
distinct peers). The only thing we need to guarantee when it comes
to unicast messages is that on a given peer they are delivered in
the order they are queued. This property we get for free from
RBTrees, so we do not need to deal with inter-message ordering
for unicast messages at all, we just have to make sure that they
are ordered correctly against multicast messages.

This allows us to reduce the write contention on the global sequence
counter a bit, by only incrementing it when allocating a new sequenc
number for a multicast message.

The new logic is:
  - The final sequence number assigned to a multicast message is a
    unique even number (as before).
  - The final sequence number assigned to a unicast message is a
    (non-unique) even number that is strictly greater than any
    sequence number thus far assigned to a multicast message and
    strictly less than any sequence number to be assigned to any
    multicast message in the future (this includes multicast
    messages that are already in flight).
  - The temporary sequence number assigned to multicast messages
    are still odd, but now they are set to be larger than any final
    sequence number that will be assigned before the next multicast
    one (this includes unicast messages that will be assigned
    after the temporary, but before the final multicast one).

The specific implementation is that we now increment the global
counter by four whenever we assigne a final multicast sequence
number. We take the current global counter + 2 as the next unicast
sequence number, and the current global counter + 3 as the
temporary multicast one.

Signed-off-by: Tom Gundersen <teg@jklm.no>