queue: fix flush operation
If we flush a peer queue, we really don't want to touch any in-flight
entries. Those are owned by the respective transaction and we better not
touch it.

Instead, iterate the rb-tree in a safe manner and drop anything that is
not in-flight and does not match the new peer ID. This is still racy, but
that's fine as we really should also drop non-matching entries during
RECV.

Signed-off-by: David Herrmann <dh.herrmann@gmail.com>