filesystem: connect - don't allocate names under the domain lock
When the domain write-lock is taken, nothing else can really be done on the bus, so we want
to be really careful about what we do with this lock taken. Adding names should be privileged
and a rare occurrence, but still, let's do the memory allocation before taking the lock.

Signed-off-by: Tom Gundersen <teg@jklm.no>