filesystem: domain - replace read locks with seqcount
This relies on an active peer (and its names) being fully
initialized. The downside is that a lookup (of id or name) may
result in a false negative (but never a false positive), in which
case a retry-loop is necessary. The chance of hitting a false
negative is very slim, so this should really only be relevant for
true negatives, which we don't really need to optimize for anyway.

As discussed with David.

Signed-off-by: Tom Gundersen <teg@jklm.no>