filesystem: domain - simplify locking during teardown
Rather than taking a domain read-lock followed by a write-lock,
simply take the write-lock once. This removes the last of the
domain read locks. Performance is not important in this path.

Deadlocks are avoided by only taking the write seqcount for the
section that was previosuly covered by the write-lock only.

Signed-off-by: Tom Gundersen <teg@jklm.no>