peer: implement tracking
Implement the TRACK and UNTRACK ioctls, including the underlying tracking
infrastructure. A new peer object is added, called bus1_peer_track. It
connects two peers, a tracker and its trackee.

Locking is a bit tricky. We now have to deal with two objects on the same
locking level, which do not have an inherent locking order. Furthermore,
both the tracker and trackee might be discoverable in arbitrary order, so
locking order (even if implied) cannot be adhered to. Therefore, we have
to do some nasty optimizations:

Inserting new track objects is easy: we pin both peers, lock them at the
same time (using the absolute address as order) and link the track object.
However, unlinking is tricky. Here we simply follow the rule: First unlink
locally (the peer you are called through / have pinned), then _try_
pinning the remote and do the same there. If that fails, we know the
remote is also about to shut down, and thus it will gc on its own. Nothing
to do for the local peer.

Signed-off-by: David Herrmann <dh.herrmann@gmail.com>