Revert "active: avoid exceeding lockdep limits"
This reverts commit 735048771b3e55c9fd33a666614d5f203b185b25. As it turns
out, lockdep is rather picky about nest-lock usage. It requires all
lock-acquisitions of the same class to be taken with the exact same
nest-lock. This is unacceptable for us, as it means we'd have to pass a
context object around with every bus1_active user.

Revert the original patch. A followup will simply provide raw bus1_active
access, which simply disables lockdep for specific call paths.