peer: drop shortcut for messages without destinations
We really should be consistent in the kernel and always do the same input
verification for all messages. This means, if a message has no
destinations, we either return an error, or we better import the message
and verify it. Otherwise, we'd be highly inconsistent.

Drop the current shortcut for n_destinations == 0 and do the proper
transaction passes.

Signed-off-by: David Herrmann <dh.herrmann@gmail.com>