pool: pre-allocate first page
The first message on any peer always needs to access at least the first
page of the destination pool. To avoid causes performance spikes if you
happen to be the first to talk to a specific peer, we now always
pre-allocate the first page on every pool. This moves the allocation
overhead into the CONNECT call.

This is not meant to be an ultimate protection of allocation overhead.
However, it avoids one of the common pitfalls in performance benchmarks.
Furthermore, every peer *always* uses its first page if it receives at
least a single message. As such we can be sure that we don't allocate
this page uselessly (a peer that doesn't receive a single message is
kinda useless / or a corner case).

Signed-off-by: David Herrmann <dh.herrmann@gmail.com>