queue: keep count of the queue size
An excessive number of messages in the queue means that the peer is
unable to process them fast enough, and that future messages will be
delivered with considerable latency.

In order to implement some sort of rate-limiting in follow-up
patches, keep count of the total number of messages in the queue.

Signed-off-by: Tom Gundersen <teg@jklm.no>