user: add quota handling
Add logic to account and check per-user, per-peer quotas. The intention
is that each peer will keep a quota object for each user that has queued
messages on the peer.

Two quotas are implemented, one for the number of queued messages and
one for the size they take up in the pool. The total number of queued
messages is soft-limited by a constant (this limit may be made more
dynamic in the future) and the memory use is naturally limited by the
size of the pool.

Note that the limit on the number of messages is not strictly enforced,
we release the lock between checking the quota and queueing the
message. The limit is arbitrary, so we don't mind going a bit over from
time to time. We may rework this in the future if we get some stricter
semantics (such as latency requirements).

In both cases, each user is given a (dynamic) quota which is given by
a constant proportion of the total that is not used by any other user.
Currently the proportion is (arbitrarily) set to 50%.

For example: if the pool size is 1024k, user A is using 128k, and user
B 256k, then user A has `(1024k - 256k) / 2 - 128k` remaining, user B
has `(1024k - 128k) / 2 - 256k` remaining, and a new user C would get
up to `(1024k - 256k - 128k) / 2`.

The aim is to minimize the reserved unused space, and the 'overcommit'
each user may have at any point in time (i.e., the amount of resources
any user use that exceed what any other user is allowed).

Signed-off-by: Tom Gundersen <teg@jklm.no>