quota: hook up quota handling
Before instantiating a message in a queue, check if the quota would
allow it. The resource usage are accounted/deaccounted where it is
consumed (in the queue and pool logic, respectively). As mentioned
in the previous commit, the pool accounting is precise, as it needs
to be as the total size of the pool is a hard limit, but the queue
accounting is not (we allow races as the limit is arbitrary anyway).

Note that the resources (pool memory usage) may not have been released
at the time the message has been dequeued, but the resource usage is
then considered the responsibility of the owner of the pool, so that is
fine.

Also note that we currently do not account for the in-flight fds, as
this needs to be just a per-user limit, and not per-pool.

Signed-off-by: Tom Gundersen <teg@jklm.no>