user: limit the inflight fds
We need to limit the number of fds a user can have inflight, as these
are otherwise not accounted on anyone (in case the sending user closes
them).

For now, follow what UDS does and simply allow the user to have as many
fds inflight as they are allowed to have open.

Note that the accounting here is done per-user, per-domain, but should
arguably be done per-user globally. We leave it like this for now as
creating multiple domains is a privileged operation, and if there are
multiple domains, one should think twice before giving the same
unprivileged user access to more than one.

Signed-off-by: Tom Gundersen <teg@jklm.no>