handle: add handle-transfer helpers
Add a new set of helpers to deal with handle transfers. They fully
implement batched pinning/creation of local handles, instantiation of
inflight handles for each destination, and the final synchronized commit
of the inflight handles.

So far the helpers are unused. Follow-up patches will hook them up in the
message transactions.

Signed-off-by: David Herrmann <dh.herrmann@gmail.com>