peer: untangle from domains
Make peers independent entities that are no longer linked to any domain.
This gets rid of a lot of code, and thus simplifies the implementation
quite a bit.

With this in place, peers are really just passive endpoints that can be
used to import handles into and then communicate. They're no longer
linked into any global context, nor discoverable or introspectable from
an independent 3rd party.

Signed-off-by: David Herrmann <dh.herrmann@gmail.com>