peer: fix race-condition on peer release
Imagine we have a peer-teardown ongoing while another operation still has
an active reference. Usually, the peer-teardown will sleep and wait until
the last active-reference-release wakes 'em up. However, imagine the
peer-teardown is not yet asleep, but the last reference is released. It
will notice that it was the last and try waking up the wake-queue. But
lets pretend it is scheduled before accessing the wait-queue, and in the
meantime the peer-teardown destroys the entire peer. Now the wait-queue
is gone and we will get an invalid memory access.

Fix this by always doing peer-releases under an rcu-read-lock. This
guarantees that the peer memory stays around and the wake-up will always
be valid.

Signed-off-by: David Herrmann <dh.herrmann@gmail.com>