bus1: inline some obvious candidates
Sadly, the kernel is still not compiled with LTO, so some of the more
obvious acquire/release fast-path helper functions will end up with
pretty costly function calls. Inline them to avoid this!

Especially in the case of active-references, we really want to avoid a
function call, if all we need is a single atomic instruction.
Furthermore, most of the heavy lifting is lockdep-only, so it is a no-op
on production builds, anyway.

Signed-off-by: David Herrmann <dh.herrmann@gmail.com>