peer: simplify CONNECT/DISCONNECT
Now that peers are completely disconnected entities, we can get rid of
most of the code-paths in CONNECT/DISCONNECT. There is no need to lock
those calls, anymore. All we need is an atomic-xchg on peer->info (since
this is nasty in combination with rcu, we borrow waitq.lock instead).

This drops most of the heavy lifting of bus1_peer. We get rid of almost
all its state and meet our ultimate goal to get everything valuable into
peer_info. All that is left in bus1_peer is "active"+"info"(which is the
reason we have the split), "waitq" (which is part of "active", we just
also borrow it for the message queue), and finally "rcu" to get delayed
releases.

Signed-off-by: David Herrmann <dh.herrmann@gmail.com>