Partially revert 55d59da046 "poll: simplify a bit"
This commit introduced a bug: It does not release the rcu-read-lock in
its early bail-out paths.

While at it, move the bus1_active_is_deactivated() check back in the rcu
read lock. While it is technically ok to call it without it, it adds 3
lines, even thought we can merge it perfectly well with the "!peer_info"
path.

Signed-off-by: David Herrmann <dh.herrmann@gmail.com>