main: drop mmap size verification
It is completely bogus to check the vma-size before calling into shmem.
We should not extend shmem with our own features, but rather let shmem
deal with the input on itself. That is, if you decide to map more than
the pool-size, you will end up with the same problems as if you map a
regular file above its size. It is not our job to protect you from doing
stupid things.

Signed-off-by: David Herrmann <dh.herrmann@gmail.com>