samples: drop unused samples infrastructure
We do not provide any samples, so drop the old infrastructure for it. Note
that we do not intend to ever add samples to the kernel module again. It
is an awful infrastructure, that is hooked into the kernel-build and is
really rather meant for kernel-code samples, not user-space. There have
been lots of problems due to it being hooked into the kernel build itself.
Hence, drop the code. If we ever want to provide samples, we do that
either as part of the tests or outside of the kernel-build-process.

Signed-off-by: David Herrmann <dh.herrmann@gmail.com>