queue: turn is_message into proper type field
Right now, we only have a single boolean on each queue-node that marks a
node as a message or not. Change this into a proper type field, that can
distinguish arbitrary numbers of types.

Follow-up patches will introduce silent messages, in which case we need a
further type, as it affects queue behavior.

Signed-off-by: David Herrmann <dh.herrmann@gmail.com>