queue: implement silent messages
This implements the BUS1_SEND_FLAG_SILENT flag. A normal message can now
be marked as silent, which will prevent the remote peer from being woken
up. The message is queued normally and can be dequeued as well. However,
if there is no non-silent message queued, it will *not* mark the FD as
readable, nor wake up the wait-queue.

This feature allows activators to queue welcome messages, which the
client will read once it starts up. The welcome message will not cause a
wake-up, though. It can also be used for other features, where the caller
sees no immediate need to wake up the client, but just wants to make sure
the message is ordered according to our global ordering rules.

Note that any inflight message is always quota accounted. So the security
is exactly the same as without this feature. You cannot cause a pool or
queue overrun with this, as you only affect your own user quota. Quotas
of other users are unaffected.

Signed-off-by: David Herrmann <dh.herrmann@gmail.com>