tests: print status message about tests
Make sure we print a message before running any tests. Otherwise, it is
annoying to figure out whether the tests were actually run, or not.
Furthermore, it allows to detect incorrect use in production systems and
may in the future tell the user why there was a delay in bootup.

Signed-off-by: David Herrmann <dh.herrmann@gmail.com>