transaction: drop duplicate-detection
This has always been a questionable feature. We needed it to avoid
hard-coded limits in the number of allowed destinations, but that is no
longer true. We now *have* to have a maximum number of handles per peer,
and this will naturally also limit the maximum number of destinations.

Drop the duplicate detection. We don't have a replacement, yet, but that
will be implemented in a follow-up.

Signed-off-by: David Herrmann <dh.herrmann@gmail.com>