peer: pin initial creds and pid_ns
We want to transmit metadata via bus messages, hence we must know the
user-ns and pid-ns of the destination we send messages to. Since our bus
is file-based, not task-based, we need to pin the user-ns and pid-ns in
the actual peer. During message transaction, we can then use it to figure
out the destination namespace.

Rather than pinning the user-ns, we pin the creds. This is a less shared
object, so we avoid modifying the user-namespace count, but still get
access to it via creds->user_ns.

For pid-namespaces, this doesn't matter much. They're pinned during each
pid-creation, anyway, so we can just pin the pid-ns directly.

Signed-off-by: David Herrmann <dh.herrmann@gmail.com>