user: split quota accounting between discharge and commit
Each message we send should atomically charge a quota, which affects both
the stats *and* global counters of the peer. If the message transaction
fails, we discharge it entirely, but if it succeeds we commit it. This
will then only discharge the in-flight data, but keep the global counters
valid until you actually release the linked objects.

This commit provides the infrastructure for it, but does not hook the
commit up. We still discharge completely even on a successful message
transaction.

Signed-off-by: David Herrmann <dh.herrmann@gmail.com>