pool/queue: silence gcc warnings
GCC complains (rightly) that we are asserting that `pool == &peer->pool`
in pool_create_for_peer(), which is tautologically true for all our
users. Simply stop passing in `pool` and drop the check. This way there
is no way to misuse the macro. Do the same for queue_init_for_peer().

Signed-off-by: Tom Gundersen <teg@jklm.no>