handle: pass initial handle in through connect
In order for peers to be able to start talking to eachother, it is necessary to
somehow pass a handle from one peer to the other without doing so through a SEND
(as without any outside handles, it is not possible to SEND between two peers).
We therefore allowed CREATE to take an existing handle in a different peer as
argument to install a copy of that handle in the local peer.

However, this sort of side-channel handle passing appears to be much more
powerful than what we need, and probably something we do not want to encourage
(it is a bit weird). It is clearly needed for bootstrap, but it does not appear
useful once setp has completed (or rather such use would be a misuse of the
concepts we want). Therefore, we restrict the api significantly, by only
allowing one outside handle to be passed in to a peer on CONNECT, and no further
side-channel handle-passing.

After all, a peer is either the root of a new bus, or added to an existing one
by having a handle to an existing peer on the bus. Setting this up at CONNECT
time makes it rather obvious what is going on.

Signed-off-by: Tom Gundersen <teg@jklm.no>