handle: distinguish between handles and nodes in the api
What we actually create/destroy are the underlying nodes (that we own), not
only their handles. So make that distinction. Otherwise it is a bit odd that
we can not perform this operation on arbitrary handles. We still do not
distinguish between handles that refer to nodes that we own and handles that
refer to nodes we do not own, so that is still something userspace would have
to get right.

Signed-off-by: Tom Gundersen <teg@jklm.no>