handles: rename bus1_handle_batch fields
A handle_entry can contain either a handle id or the pointer to a pinned
handle, we account for these two kinds of representation in handle_batch.
Make this a bit more obvious by renaming the fields.

@n_entries count the total number of ids or handles (but excludes the
.next pointers, so think of it as the number of entries if the batch had
been a linear array). @n_handles count the number of true (pinned)
handles.

Signed-off-by: Tom Gundersen <teg@jklm.no>