tests: rework tests to use new bus1_client API
This imports the new bus1_client API and reworks the tests to use them.
Since we cannot use the b1_* prefix (reserved for libbus1), we use the
kernel prefix bus1_* but limit ourselves to bus1_client_*.

The b1_test infrastructure is renamed to just test_* to distinguish it
from the in-kernel tests.

A basic test-api test is added, which should be the first test that is
run and checks for very basic module functionality, that is needed for
any other test to succeed.

Signed-off-by: David Herrmann <dh.herrmann@gmail.com>