peer: fix missing module-reference after alloc_file()
Unfortunately, alloc_file() consumes the reference of the caller via
f_ops. That is, we're required to provide fops_get() to them. This in
turn is just a public wrapper around try_module_get(), and since we
already hold a reference in this case, we can safely use __module_get().

This fixes a negative module-refcount after clone operations.

Signed-off-by: David Herrmann <dh.herrmann@gmail.com>