handle: fix NULL-deref on attach
When attaching an owner handle, then handle->node->owner.holder is
obviously NULL (as we're about to attach it). We already check for this,
but we fail to properly acquire the fallback pointer (which is the peer
you want to attach).

Fix the NULL-deref by setting 'owner' properly in case of owner attach.

Signed-off-by: David Herrmann <dh.herrmann@gmail.com>