handle: document that rb-nodes might be stray
During peer reset (and shutdown) we flush the entire handle-tree and
release/destroy them unlocked. This requires to keep the tree intact so
we can traverse it. We could, technically, reset the nodes afterwards,
but it would require us to lock the handle again. We avoid this, as such
the rb-nodes might contain stray pointers.

Document this behavior so it is obvious.

Signed-off-by: David Herrmann <dh.herrmann@gmail.com>