peer: copy handles into slice on receive
Just like for file-descriptors, copy over the handle-IDs into the slice
before returning a message via RECV.

Note that we explicitly do this on RECV, rather than on SEND. We must do
this *after* we know the final timestamp of a transaction, at which point
we are not allowed to fail. However, the copy-operation might fail, so we
must instead do the copy on RECV, where we can still bail out if it
fails.

Signed-off-by: David Herrmann <dh.herrmann@gmail.com>