transaction: refuse message if destination is destroyed
Node destruction can always race message transactions. As such, we must
make sure to refuse a transaction that happens to be ordered *after* the
destruction of its destination.

Use the new bus1_handle_get_inorder_id() to order transactions against
possible destruction of the destination, and bail out if required.

Signed-off-by: David Herrmann <dh.herrmann@gmail.com>