handle: use atomic_inc_return() on conflicts directly
Rather than using bus1_handle_acquire() and expecting it to not fail, use
atomic_inc() directly as we know we don't need the error handling of
bus1_handle_acquire() (and we definitely don't care for the special owner
handling).

Make the code simpler by not relying on the public user-ref to
inflight-ref converter, but directly access inflight refs.

Signed-off-by: David Herrmann <dh.herrmann@gmail.com>