util: fix bus1_clone_file() to take cdev reference
Unfortunately, alloc_file() is not a proper counter-part for fput(), so
we have to take care of external references ourself. In particular, we
need to take cdev references if we clone a cdev.

To prevent similar bugs in the future, we now prevent to clone anything
but bus1-files. If we need this for more file types, we can add them
later on.

Signed-off-by: David Herrmann <dh.herrmann@gmail.com>