handle: reject handle operations if no user refs exist
If a user performs a handle operation, make sure to reject it if they do
not own a user reference. This is particularly important for node owners,
as they can always lookup the node, but should not be able to use it if
they do not own a user reference.

Note that the check for n_user is completely unlocked. However, we always
follow with a acquire operation of an inflight reference. If n_user jumps
between the check and acquisition, we don't care. If the acquisition
succeeds, we just assume it succeeded atomically with the n_user check.
The only side-effect that might have a real effect, is if inflight refs
drop to 0 in between. However, in that case, the acquisition fails and we
bail out anyway.

Hence, just do the n_user check before the acquisition and we are good to
go.

Signed-off-by: David Herrmann <dh.herrmann@gmail.com>