handle: rewrite handle management
This rewrites most of the handle management, so we can finally implement
the remaning pieces, including notification support. Changes include:

 - keep bus1_handle completely intern to handle.c, only export the batch
   helpers, and destination management

 - use bus1_node as top-level object, but still hidden in handle.c

 - return node-handle via CLONE, so user-space does not have to guess it

 - drop unused 'handle' argument from RESET

The big part of this commit is a complete rewrite of the handle
management to allow proper async notification support, synchronous peer
shutdown, and simplify the logic quite a bit. Furthermore, handle IDs are
no longer re-used. User-space now has full control over its IDs and the
kernel-ioctl entries serve as barriers for user-space to rely on.

Signed-off-by: David Herrmann <dh.herrmann@gmail.com>