transaction: properly implement BUS1_SEND_FLAG_CONTINUE
The purpose of BUS1_SEND_FLAG_CONTINUE is to prevent a destination from
tampering with the message transaction targetted at it. That is, whenever
you send a message with BUS1_SEND_FLAG_CONTINUE, the transaction is
supposed to fail only if you yourself caused the error. A possible
destination of the transaction must not be able to cause the transaction
to fail. This is very important for signal-based transactions, where the
sender usually never trusts the recipients, but wants to merge multiple
transactions into a single multicast for performance and ordering
reasons.

Now the only reason a destination can cause a message transaction to fail
is if the destination quota exceeds, or the destination plainly does not
exist, anymore. Therefore, rather than ignore *all* errors, this patch
fixes the transaction to only ignore errors from bus1_message_allocate().
That is, only if the target message cannot be allocated, we ignore the
error. Furthermore, we now still carry the defunct message stub around,
to make sure we know about failed messages at commit-time. This
guarantees that we can actually notify the failed destination about this
error (via the n_dropped counter), in case this was not a malicious
attempt to DoS the sender (which might be completely reasonable in a
normal system).

While at it, this also completes the destination-ID-allocation feature
and properly copies the new IDs to user-space. Now that messages are
properly tracked, we get this feature basically for free. We simply have
to carry a "userid" pointer in a message, in case a copy is desired.

Signed-off-by: David Herrmann <dh.herrmann@gmail.com>