tree-wide: don't overload EINVAL
EINVAL indicates a programming error in userspace, however, we also used it for
kernel programming errors, which are also goarded with WARN_ON. Use a distinct
error code for this to make debugging a bit simpler. This is for all intents and
purposes dead code, and should never be encountered from userspace, so don't
document it.

Signed-off-by: Tom Gundersen <teg@jklm.no>