handle: tick late on node-destruction
When a node-destruction is done, we always sync the owning queue while we
commit the destruction timestamp. This is important so we can reliable
return error codes on racing messages targetted at exactly this node.
However, we must also make sure to *TICK* the queue, or we might end up
with multiple messages with the same timestamp, which ends up being
unrecoverable.

Signed-off-by: David Herrmann <dh.herrmann@gmail.com>