quotas: rework quota to have per-user as well as per-peer limits
The new logic is:

Every user has a global limit on each of the possible resources it
can consume in the kernel. Moreover, every peer has a local limit
on how much it may consume of the same resources.

How much of a given resource is available to a given peer, is
capped by how much of that resource remains available to the user
owning the peer.

The available resources are distributed among each of the sending
users in the same fashion as before: each user has available half
of the total resources not used by anybody else.

The global limits on the number of messages, the number of handles
and the number of fds are static. We do not enforce a global limit
on the pool usage, but let shmem handle this for us.

The local limit on the number of messages and the number of handles
is so far unbounded, in the future it will be possible to specify it
in the ioctl when initializing a peer. The local limit on the pool
usage is the pool-size specified in the ioctl when initializing the
peer. The local limit on the number of fds is the RLIMIT_NOFILE of
the process that initialized the peer at the time of initialization.

Signed-off-by: Tom Gundersen <teg@jklm.no>