handle: make n_user a proper subordinate of n_inflight
This changes n_user of bus1_handle to be a proper subordinate of
n_inflight. Meaning that n_inflight is the major factor controlling the
lifetime of handles. n_user is just the user-visible/controlled part of
it, which no longer has any runtime control over the handle, apart from
pinning a single inflight ref.

This has the major benefit that handle lifetime can no longer be skipped
early. Instead, the handle lifetime is now fully synced with the
n_inflight pinning. This simplifies the logic *a lot*, as we can now move
*all* destruction handling of non-owner handles to the last inflight ref.

The only exception is owner-handles, which must be pinned until the node
is fully destroyed. However, this is now also linked to n_inflight, but in
combination with handle destruction.

While at it, this commit also introduces bus1_handle_dest, which wraps
destination handling. This allows us to fully hide handles in handle.c,
and the only "public" objects are dests, transfers, and inflights.

Signed-off-by: David Herrmann <dh.herrmann@gmail.com>